Refactor snackbar store function declaration for a more concise and better pattern and also return ErrorDetails in Auth Controller layer
