+++ lineshift-v3-client/src/services/sports.api.js
@@ -1,17 +1,15 @@
-import { useSnackbarStore } from '@/stores/snackbar.store';
 import axiosInstance from './instance/api';
 import { sportsEndpoint } from './instance/endpoints';
+import parseNetworkErrorResponse from './parsers/api.response.parser';
 
 
 const sportsApi = {
   async getAllSports() {
-    const { showError } = useSnackbarStore();
     try {
       const response = await axiosInstance.get(sportsEndpoint);
       return response.data;
     } catch (error) {
-      showError(`Error fetching all sports: ${error}`);
-      throw error;
+      throw parseNetworkErrorResponse(error);
     }
   },
 };
