+++ lineshift-v3-client/src/stores/resources/sports.store.js
@@ -1,11 +1,10 @@
 import sportsApi from '@/services/sports.api';
 import { defineStore } from 'pinia';
 import { ref } from 'vue';
-import { useSnackbarStore } from './snackbar.store';
+import { useSnackbarStore } from '../snackbar.store';
 
 export const useSportsStore = defineStore('sports', () => {
-  const snackbarStore = useSnackbarStore();
-  const { showSuccess, showError } = snackbarStore;
+  const {showError} = useSnackbarStore()
 
   // State
   const sports = ref([]);
@@ -19,8 +18,7 @@ export const useSportsStore = defineStore('sports', () => {
     }
     catch (error) {
       console.error(error);
-      showError(error);
-      throw error;
+      showError(error.message);
     }
   };
